+++ kirksdk/account_api.go
@@ -80,6 +80,39 @@ type AccountClient interface {
 
 	// GetQcosClient 用于得到与某个 App 交互的 QcosClient
 	GetQcosClient(ctx context.Context, appURI string) (client QcosClient, err error)
+
+	// CreateAppGrant 将应用授权给用户
+	CreateAppGrant(ctx context.Context, appURI, username string) (err error)
+
+	// DeleteAppGrant 删除应用授权
+	DeleteAppGrant(ctx context.Context, appURI, username string) (err error)
+
+	// ListAppGrantedUsers 列出应用已授权的用户列表
+	ListAppGrantedUsers(ctx context.Context, appURI string) (ret []AppGrantedUser, err error)
+
+	// ListGrantedApps 列出已被授权的应用
+	ListGrantedApps(ctx context.Context) (ret []AppInfo, err error)
+
+	// GetGrantedAppKey 获取被授权应用的key
+	GetGrantedAppKey(ctx context.Context, appURI string) (ret GrantedAppKey, err error)
+
+	// GetAppspecs 获得应用模板信息
+	GetAppspecs(ctx context.Context, specURI string) (ret SpecInfo, err error)
+
+	// ListPublicspecs 列出公开应用的模板
+	ListPublicspecs(ctx context.Context) (ret []SpecInfo, err error)
+
+	// ListGrantedspecs 列出被授权应用的模板
+	ListGrantedspecs(ctx context.Context) (ret []SpecInfo, err error)
+
+	// GetVendorManagedAppStatus 获得VendorManaged应用运行状态
+	GetVendorManagedAppStatus(ctx context.Context, appURI string) (ret VendorManagedAppStatus, err error)
+
+	// GetVendorManagedAppEntry 获得VendorManaged应用入口地址
+	GetVendorManagedAppEntry(ctx context.Context, appURI string) (ret VendorManagedAppEntry, err error)
+
+	// VendorManagedAppRepair 尝试修复VendorManaged应用
+	VendorManagedAppRepair(ctx context.Context, appURI string) (err error)
 }
 
 // AccountConfig 包含创建 AccountClient 所需的信息
@@ -177,4 +210,43 @@ type UpdateAlertMethodArgs struct {
 	Mobile      string `json:"mobile"`
 	Nationality string `json:"nationality"`
 	Code        string `json:"code"`
+}
+
+// AppGrantedUser 包含列出应用被授权的用户信息
+type AppGrantedUser struct {
+	ID   uint32 `json:"id"`
+	Name string `json:"name"`
+}
+
+// GrantedAppKey 包含被授权应用的key信息
+type GrantedAppKey struct {
+	Ak string `json:"ak"`
+	Sk string `json:"sk"`
+}
+
+// SpecInfo 包含 Spec 的相关信息
+type SpecInfo struct {
+	URI       string    `json:"uri"`
+	Owner     string    `json:"owner"`
+	Title     string    `json:"title"`
+	Ver       uint32    `json:"ver"`
+	Verstr    string    `json:"verstr"`
+	Desc      string    `json:"desc,omitempty"`
+	Brief     string    `json:"brief"`
+	Icon      string    `json:"icon"`
+	Seedimg   string    `json:"seedimg"`
+	Entryport uint16    `json:"entryport"`
+	Ctime     time.Time `json:"ctime"`
+	Mtime     time.Time `json:"mtime"`
+}
+
+// VendorManagedAppStatus 包含应用运行状态信息
+type VendorManagedAppStatus struct {
+	Status  string `json:"status"`
+	Message string `json:"message"`
+}
+
+// VendorManagedAppEntry 包含应用入口地址
+type VendorManagedAppEntry struct {
+	Entry string `json:"entry"`
 }