+++ kirksdk/kirksdk.go
@@ -1,3 +1,3 @@
 package kirksdk
 
-const Version = "1.1.0"
+const Version = "2.1.0"