+++ kirksdk/qcos_client_test.go
@@ -248,6 +248,18 @@ func TestServicesInspect(t *testing.T) {
 				Name:      "v2",
 			},
 		},
+		ApPorts: []ServiceApPort{
+			ServiceApPort{
+				ApID:         "1000001",
+				Type:         "DOMAIN",
+				Domain:       "abcd1234",
+				UserDomains:  []string{"www.aa.com"},
+				FrontendPort: "80",
+				BackendPort:  "8080",
+				Proto:        "HTTP",
+				Enabled:      true,
+			},
+		},
 	}
 	ret := `{
     "containerIps": [
@@ -350,7 +362,17 @@ func TestServicesInspect(t *testing.T) {
             "name": "v2",
 			"unitType": "SSD1_16G"
         }
-    ]
+    ],
+    "apPorts": [{
+    	"apId": "1000001",
+    	"type": "DOMAIN",
+    	"domain": "abcd1234",
+    	"userDomains": ["www.aa.com"],
+    	"frontendPort": "80",
+    	"backendPort": "8080",
+    	"proto": "HTTP",
+    	"enabled": true
+    	}]
 }`
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		assert.Equal(t, expectedUrl, r.URL.Path)
@@ -622,4 +644,37 @@ func TestUpdateJob(t *testing.T) {
 	}
 	err := client.UpdateJob(context.TODO(), "default", args)
 	assert.NoError(t, err)
+}
+
+func TestGetWebproxy(t *testing.T) {
+	expectedUrl := "/v3/webproxy"
+	expectedMethod := "POST"
+	expectedBody := `{"backend":"10.128.0.1:8080"}`
+	returnBody := `{
+    "backend": "10.128.0.1:8080",
+    "oneTimeUrl": "http://dummy.com"
+}`
+
+	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		assert.Equal(t, expectedUrl, r.URL.Path)
+		assert.Equal(t, expectedMethod, r.Method)
+		body, err := ioutil.ReadAll(r.Body)
+		assert.NoError(t, err)
+		assert.Equal(t, expectedBody, string(body))
+		fmt.Fprintln(w, returnBody)
+	}))
+	defer ts.Close()
+
+	client := NewQcosClient(QcosConfig{
+		Host: ts.URL,
+	})
+
+	args := GetWebProxyArgs{
+		Backend: "10.128.0.1:8080",
+	}
+
+	webproxy, err := client.GetWebProxy(context.TODO(), args)
+	assert.NoError(t, err)
+	assert.Equal(t, "10.128.0.1:8080", webproxy.Backend)
+	assert.Equal(t, "http://dummy.com", webproxy.OneTimeURL)
 }